pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drama-finder"
version = "0.1.0"
description = "Chinese drama screenshot localization system backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16.0",
  "fastapi>=0.116.0",
  "google-genai>=1.30.0",
  "pgvector>=0.4.1",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.10.0",
  "pyyaml>=6.0.2",
  "redis>=6.4.0",
  "rq>=2.6.0",
  "sqlalchemy>=2.0.43",
  "uvicorn[standard]>=0.35.0",
  "python-multipart>=0.0.20",
]

[project.optional-dependencies]
pipeline = [
  "faster-whisper>=1.2.0",
  "numpy>=2.3.0",
  "scenedetect>=0.6.5.2",
]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
  "ruff>=0.12.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
